Simplify pipeline interface (quater)
