+++ myia/pipeline/standard.py
@@ -3,48 +3,23 @@
 
 from ..abstract import Context, abstract_inferrer_constructors
 from ..ir import GraphManager
-from ..pipeline.resources import (
-    BackendResource,
-    ConverterResource,
-    DebugVMResource,
-    InferenceResource,
-    scalar_object_map,
-    standard_method_map,
-    standard_object_map,
-)
 from ..prim import py_registry, vm_registry
-from ..utils import Partial, Partializable
 from ..validate import (
     validate_abstract as default_validate_abstract,
     whitelist as default_whitelist,
 )
 from . import steps
 from .pipeline import PipelineDefinition
-
-
-class Resources(Partializable):
-    def __init__(self, **members):
-        self._members = members
-        self._inst = {}
-
-    def __getattr__(self, attr):
-        if attr in self._inst:
-            return self._inst[attr]
-
-        if attr in self._members:
-            inst = self._members[attr]
-            if isinstance(inst, Partial):
-                try:
-                    inst = inst.partial(resources=self)
-                except TypeError:
-                    pass
-                inst = inst()
-            self._inst[attr] = inst
-            return inst
-
-    def __call__(self):
-        return {'resources': self}
-
+from .resources import (
+    BackendResource,
+    ConverterResource,
+    DebugVMResource,
+    InferenceResource,
+    Resources,
+    scalar_object_map,
+    standard_method_map,
+    standard_object_map,
+)
 
 standard_resources = Resources.partial(
     manager=GraphManager.partial(),
@@ -73,22 +48,18 @@ standard_resources = Resources.partial(
 
 
 standard_pipeline = PipelineDefinition(
-    # resources=standard_resources,
-    resources=Resources,
-    steps=dict(
-        RRR=standard_resources,
-        parse=steps.step_parse,
-        resolve=steps.step_resolve,
-        infer=steps.step_infer,
-        specialize=steps.step_specialize,
-        simplify_types=steps.step_simplify_types,
-        opt=steps.step_opt,
-        opt2=steps.step_opt2,
-        cconv=steps.step_cconv,
-        validate=steps.step_validate,
-        compile=steps.step_compile,
-        wrap=steps.step_wrap,
-    )
+    RRR=standard_resources,
+    parse=steps.step_parse,
+    resolve=steps.step_resolve,
+    infer=steps.step_infer,
+    specialize=steps.step_specialize,
+    simplify_types=steps.step_simplify_types,
+    opt=steps.step_opt,
+    opt2=steps.step_opt2,
+    cconv=steps.step_cconv,
+    validate=steps.step_validate,
+    compile=steps.step_compile,
+    wrap=steps.step_wrap,
 )
 
 
@@ -98,22 +69,18 @@ scalar_pipeline = standard_pipeline.configure({
 
 
 standard_debug_pipeline = PipelineDefinition(
-    # resources=standard_resources,
-    resources=Resources,
-    steps=dict(
-        RRR=standard_resources,
-        parse=steps.step_parse,
-        resolve=steps.step_resolve,
-        infer=steps.step_infer,
-        specialize=steps.step_specialize,
-        simplify_types=steps.step_simplify_types,
-        opt=steps.step_opt,
-        opt2=steps.step_opt2,
-        cconv=steps.step_cconv,
-        validate=steps.step_validate,
-        export=steps.step_debug_export,
-        wrap=steps.step_wrap,
-    )
+    RRR=standard_resources,
+    parse=steps.step_parse,
+    resolve=steps.step_resolve,
+    infer=steps.step_infer,
+    specialize=steps.step_specialize,
+    simplify_types=steps.step_simplify_types,
+    opt=steps.step_opt,
+    opt2=steps.step_opt2,
+    cconv=steps.step_cconv,
+    validate=steps.step_validate,
+    export=steps.step_debug_export,
+    wrap=steps.step_wrap,
 ).configure({
     'RRR.backend.name': False
 })
