+++ tests/test_grad.py
@@ -84,17 +84,14 @@ def grad_wrap(graph, argspec):
 
 
 grad_pipeline = PipelineDefinition(
-    resources=standard_resources,
-    steps=dict(
-        RRR=standard_resources,
-        parse=steps.step_parse,
-        resolve=steps.step_resolve,
-        infer=steps.step_infer,
-        specialize=steps.step_specialize,
-        opt=steps.step_debug_opt,
-        validate=steps.step_validate,
-        export=steps.step_debug_export,
-    )
+    RRR=standard_resources,
+    parse=steps.step_parse,
+    resolve=steps.step_resolve,
+    infer=steps.step_infer,
+    specialize=steps.step_specialize,
+    opt=steps.step_debug_opt,
+    validate=steps.step_validate,
+    export=steps.step_debug_export,
 ).configure({
     'RRR.operation_whitelist': grad_whitelist,
     'RRR.validate_abstract': grad_validate_abstract
